+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Canonical reference for changes, improvements, and bugfixes for mql.
 
 ## Next
+* feat: add support for table column mapping by @dlclark in [[PNR](https://github.com/hashicorp/mql/pull/45)]
 
 ## 0.1.4 (2024/05/14)